complete rework of 900x bootloader/reflash code, bugfixs
Prior to this point, a 900x modem that is stuck in "solid red light" bootloader mode will be permanently stuck there.  
This commit allows the radio to be recovered from that mode by trying to interact with the modem at multipkme baudrates, including the same baudrate as the esp8266 bootloader (74880), and the stock baudrate (57600). 
This is also helped by the fact that we  now try to do a BOOTNEW at the weird baudrate, causing a soft reboot of the 900x, and then drop back to baudrate of 57600 for  aretry of the upload/flash stage.  
THis code is a bit poorly laidout, and includes the occassional goto for triggering retries under odd error conditions, and at least one new global variable, and a number of short-circuit tests that speed-up the entire precess during error conditions,  but is proven functional and tested repeatedly on real hardware.