Remap RSSI-to-percent function to be more useful
This is a bit silly, ugly, and hacky. This makes the first 10% of the range more useful. 206 or better shows as 99%, 94 shows as 90%. Below that, it works like the old calculation did (except it goes from 0 to 90 over that range instead of 0 to 99).

The old method considered signal levels to be 99% until way too late, where it would suddenly drop off. This gives slightly more info before the sharp drop.